backend-api/requirements.txt
Flask
Flask-Cors
pymongo
python-dotenv
APScheduler
pyJWT
gunicorn